# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Predict API",
    "description": "Satellite orbit prediction",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "station.local:3000",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/satellites": {
      "get": {
        "summary": "Get a list of satellites in the predict database.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/satellites/{name}": {
      "get": {
        "summary": "Get information about a specific satellite.",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "type": "string",
            "in": "path",
            "description": "The name of the satellite."
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "satellites/{name}/doppler": {
      "get": {
        "summary": "Get dopper information about a specific satellite.",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "type": "string",
            "in": "path",
            "description": "The name of the satellite."
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/satellites/{name}/tle": {
      "get": {
        "summary": "Get the Keplerian information (tle) about a specific satellite.",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "type": "string",
            "in": "path",
            "description": "The name of the satellite."
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/satellites/{name}/position": {
      "get": {
        "summary": "Get current position information about a specific satellite.",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "type": "string",
            "in": "path",
            "description": "The name of the satellite."
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/satellites/{name}/predict": {
      "get": {
        "summary": "Get orbit prediction for a specific satellite.",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "type": "string",
            "in": "path",
            "description": "The name of the satellite."
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/satellites/{name}/predict/{start}": {
      "get": {
        "summary": "Get orbit prediction for a specific satellite.",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "type": "string",
            "in": "path",
            "description": "The name of the satellite."
          },
          {
            "name": "start",
            "required": true,
            "type": "string",
            "in": "path",
            "description": "The start time for the requested prediction."
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/sys": {
      "get": {
        "summary": "Get location data.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/sys/version": {
      "get": {
        "summary": "Get the version of predict.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/sys/mode": {
      "get": {
        "summary": "Get current mode of predict.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/sys/time": {
      "get": {
        "summary": "Get the system time.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/sys/tle": {
      "get": {
        "summary": "Reload the TLE files.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ephemeris/sun": {
      "get": {
        "summary": "Get ephemeris information about the sun.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ephemeris/moon": {
      "get": {
        "summary": "Get ephemeris information about the moon.",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  }
}